refactor: rename repo to repo_jdbc
- rename repo to repo_jdbc
